backend/requirements.txt
torch 
torchvision 
torchaudio
transformer
accelerate
sentencepiece